requirements_v1.txt
pandas
matplotlib
psycopg2-binary
seaborn
openpyxl
requests
python-dotenv
msgraph-sdk
pyxlsb
oauth2client
s3fs
yfinance
tabulate
statsmodels
google-auth
google-api-python-client